Use uniform prior for indel because a transformation of uniform->non-uniform->uniform was applied.

+++ consensus.hpp
@@ -3049,9 +3049,9 @@ appendVcfRecord(std::string & out_string, std::string & out_string_pass, VcStats
         std::array<double, N_MODELS> testquals = {0};
         unsigned int tqi = 0;
         // const double a_nogerm_q = (double)(n_nogerm_q + 0.0*MIN(MAX(0, t_nogerm_q),25));
-        double t2n_finq  = max_min01_sub02(MIN(t2n_rawq           , t2t_powq           )    , MIN(t2n_rawq, t2n_powq), t2n_contam_q);
+        double t2n_finq  = max_min01_sub02(MIN(t2n_rawq           , t2t_powq        )              , MIN(t2n_rawq, t2n_powq), t2n_contam_q);
         // 0 // n_nogerm_q and t2n_powq should have already bee normalized with contam
-        testquals[tqi++] = max_min01_sub02(MIN(tn_trawq, tn_tpowq + indel_ic) + 1.0*t2n_finq, a_nogerm_q,              t2n_contam_q) + 0.0*t2n_finq + indel_aq;
+        testquals[tqi++] = max_min01_sub02(MIN(tn_trawq, tn_tpowq + (double)indel_ic) + 1.0*t2n_finq, (double)a_nogerm_q,     t2n_contam_q) + 0.0*t2n_finq + indel_aq;
         //testquals[tqi++] = MIN(tn_trawq, tn_tpowq + tvn_powq) - MIN(tn_nrawq, MAX(0.0, tn_npowq - tvn_or_q));
         testquals[tqi++] = MIN(tn_trawq, tvn_rawq * 2 + 30);
         testquals[tqi++] = MIN(tn_trawq, tvn_rawq     + tn_tpowq);
